Move enum GameState
